chore: add useful methods for handling the data (not complete yet)

+++ engines/knowledge-engine/src/objects/edge.rs
@@ -1,15 +1,15 @@
 use crate::enumerations::Relationship;
 
 /// Main struct used for the connections between the nodes on the graph
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, Ord, PartialOrd, Eq, PartialEq)]
 pub struct Edge{
-    weight: f64,
+    weight: i32,
     relationship: Relationship
 }
 
 impl Edge{
     /// Constructor with the required fields to create an Edge
-    pub fn new(weight: f64, relationship: Relationship) -> Self {
+    pub fn new(weight: i32, relationship: Relationship) -> Self {
         Self { weight, relationship }
     }
 }
